PLEASE READ DESCRIPTION
This process is very fast but does use quite a bit of memory since I haven't implemented chunking yet. That is why there is a maxPixels variable that will resize the image if it exceeds the maximum allowed pixels. I suggest you put this at somewhere between 1000x1000 and 2000x2000 for now. Also, when displaying the image to the client, there is no use in using the png, just convert the Base64 back to Webp and display that.